feat(conversation):WB-3957- empty trash (#647)
* feat(conversation):WB-3957- empty trash

* feat(conversation):WB-3957 clean reload List component with key

* refactor(conversation):WB-3957 create useToolbarActions hook

* refactor(conversation):WB-3957 rename MessageList.tsx

* refactor(conversation):WB-3957 create MessageItem component

* fix(conversation):WB-3957 manage List reloading for update global checkbox

* fix(conversation):WB-3957 fix List on frontend framwork to clean hack in MessageList

* refactor(conversation):WB-3957 add useToolbarVisibility hook

* refactor(conversation):WB-3957 reload list when searchparams change

* fix(conversation): #WB-3444 keep id after create

* feat(conversation): #WB-3925, add or remove attachments to draft messages (#644)

* feat(conversation): #WB-3925, add attachments to draft messages

* service testing

* wip

* wip

* todo: some renders fail

* tried fixing previous bug

* bug fixed

* Added mutating feedback to the user

* feat: Remove all attachments

* fix tests

* todo: attach files on a new draft do not render well

---------
