+++ conversation/frontend/src/features/message-list/hooks/useSelectedMessages.ts
@@ -1,0 +1,7 @@
+import { MessageMetadata } from '~/models';
+import { useSelectedMessageIds } from '~/store';
+
+export default function useSelectedMessages(messages: MessageMetadata[]) {
+  const selectedIds = useSelectedMessageIds();
+  return messages?.filter((message) => selectedIds.includes(message.id)) || [];
+}
